+++ HISTORY.md
@@ -19,13 +19,40 @@
 
 Developers are not forced to upgrade if they don't really need it. Upgrade whenever you feel ready.
 
-**How to upgrade**: Open your command-line and execute this command: `go get github.com/kataras/iris@v11.2.0`.
+**How to upgrade**: Open your command-line and execute this command: `go get github.com/kataras/iris@master`.
 
-# We, 24 July 2019 | v11.2.1
+# Tu, 30 July 2019 | v11.2.3
+
+TODO:
+
+- Different parameter types in the same path (done).
+- [Content negotiation](https://developer.mozilla.org/en-US/docs/Web/HTTP/Content_negotiation) (in-progress)
+
+# We, 24 July 2019 | v11.2.2
+
+Sessions as middleware:
+
+```go
+import "github.com/kataras/iris/sessions"
+// [...]
+
+app := iris.New()
+sess := sessions.New(sessions.Config{...})
+
+app.Get("/path", func(ctx iris.Context){
+    session := sessions.Get(ctx)
+    // [work with session...]
+})
+```
+
+- Add `Session.Len() int` to return the total number of stored values/entries.
+- Make `Context.HTML` and `Context.Text` to accept an optional, variadic, `args ...interface{}` input arg(s) too.
+
+## v11.1.1
 
 - https://github.com/kataras/iris/issues/1298
 - https://github.com/kataras/iris/issues/1207
 
 # Tu, 23 July 2019 | v11.2.0
 
-Read about the new release at: https://dev.to/kataras/iris-version-11-2-released-22bc
+Read about the new release at: https://www.facebook.com/iris.framework/posts/3276606095684693